# Zen4: Faster PP for IQ2_KS, IQ4_KS, IQ5_KS